Keep two lines for basePath. Added space to a print.
There is a problem with hardcoding basePath: It is machine dependent. I need a way to automatically define it based on the local machine; for now, I just hardcoded both and will comment out the unused one. Also added a space to make the print statement before read-in look nicer.