Renamed set_basepath to set_paths and updated return value
set_paths returns two values: the first (index 0) is the full path for all simulation runs; the second (value 1) is the relative path, within the first path, that contains data. On SCC and Stampede, the two paths are the same, but on other machines, there may be a subpath dedicated to holding data.